% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credibleInterval.R
\name{credibleInterval}
\alias{credibleInterval}
\title{Credible Interval}
\usage{
credibleInterval(posteriorSamples, width)
}
\arguments{
\item{posteriorSamples}{Posterior distribution samples}

\item{width}{CI width}

\item{digits}{digits for the CI}
}
\value{
Credible interval
}

