% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybayesian.R
\docType{package}
\name{easybayesian}
\alias{easybayesian}
\alias{easybayesian-package}
\title{easybayesian: A package that makes simple bayesian estimation easy}
\description{
The easybayesian package provides three categories of important functions:
\code{\link{bayesiandashboard}}, \code{\link{stanlm}}, 
\code{\link{regtbl}}, \code{\link{posteriorplot}}, 
\code{\link{interpret}}, and \code{\link{updateci}}
}

